Create MIQROOT/vmdb/vendor/assets symlink and allow Apache to follow
the vmdb/vendor/assets directory is used to store compiled assets for the UI.
bin/update via the bower command tries to change ownership on some of these
files but while running from mapped source this will fail.
Create a symlink to the local appliance running on the VM under one's MAC
so the assets are stored locally.

In addition add to the Apache configuration so that symbolic links are followed
properly at all levels.